refactor: Remove static keyword from PlayBgm and PlayUiClick methods
The AudioManager class had two public methods, PlayBgm and PlayUiClick, that were marked as static. This commit removes the static keyword from these methods to allow them to be called on an instance of the AudioManager class. Additionally, a new optional parameter was added to the private method PlayAudio to allow for playing audio clips as one-shots. Finally, a TODO comment was added to address a memory leak issue in the code.